Docstrings, rename classes to include Type suffix
This commit adds docstrings to the classes in types.py and also
directly names the classes with the Type suffix. For example,
Boolean -> BooleanType. The instantiations at the bottom of the
file have been removed.

One additional thing included is an update to the test_types.py
to instantiate the types. For example, anywhere `BooleanType` was
used, it's been changed to `BooleanType()` and comparisons to it
have been changed from `==` to `isinstance`.