Spark: Fix Rewrite Overlap Test Function
Previously the overlap function would not correctly find overlaps when the bounds for
overlapping files were equivelent, or if a single file was returned. The new function
correctly only ignores files that are exactly the same (path and bounds) and throws an
error if overlaps are being checked for on a single file.