Nessie: simplify code in Nessie catalog
This adds a new `NessieIcebergClient` class that encapsulates some of
the features that have been used across the `NessieCatalog`.
It also slightly improves the error messages to include the reference or
the table name when something fails.

Note that the `testFailure()` method was updated, because we're now
always using the same `UpdateableReference` between `NessieCatalog` and
`NessieTableOperations`, so that it's higly unlikely that those two get
out-of-sync.