Flink: fix out of order appends in upsert UT
In Flink, only VALUES with literal tuples can be converted to a single
LogicalValues node. VALUES with other type of tuples will be converted
to a chain with multiple LogicalProject nodes, like following:

    LogicalValues -> (# number or rows)LogicalProject -> LogicalUnion.

The `TO_DATE('xxxx-xx-xx')` expression is not a literal node, so the
VALUES used in upsert test cases will be converted in the latter way.
The order of tuples can not be guaranteed, which can lead to unstable
test results.

Changing `TO_DATE('xxxx-xx-xx')` to `DATE 'xxxx-xx-xx'` can make it a
literal node and ensure that the VALUES expression can be converted to a
single LogicalValues node.