Python: Set Python 3.8 as minimal version
Python 3.8 has some nice features, like:

- The Walrus operator := https://realpython.com/python-walrus-operator/
- Ability to cache properties: @cached_property https://docs.python.org/3/library/functools.html#functools.cached_property

3.7 is a LTS that has still support for a bit over a year: https://endoflife.date/python

I've added 3.10 to the CI as well. Not sure if we can deprecate 3.7 just yet, since it is still being used a lot:

That being said, upgrading from 3.7 to 3.8 should be quite easy now a days with minimal breaking changes on the Python side, and just swapping out a base docker image.