Fix invalid Stream to Iterable conversion in decryption
`Stream.iterator` is a one-shot operation, cannot be called multiple
times, so `stream::iterator` shouldn't be used as an implementation of
`Iterable`.

While we could easily change the `InputFilesDecryptor` to implement
`Iterable` correctly, it wouldn't convey the intention that the data is
read at most once. To convey that, the encryption API is changed from
accepting `Iterable` to accepting `Iterator`.