OpenAPI: Move to oneOf for the types
All the types match exactly one of the schemas:

```json
"string"

{"type": "struct", ...}

{"type": "list", ...}

{"type": "map", ...}
```

For more information:
https://swagger.io/docs/specification/data-models/oneof-anyof-allof-not/

We could also add a discriminator if we bump the primitive type
to an object as well. My suggestion would be to break down:

```json
"fixed[22]"
```

into

```json
{"type": "fixed", "length": 22}
```

This would allow us to set:

```yaml
    Type:
      oneOf:
        - $ref: '#/components/schemas/PrimitiveType'
        - $ref: '#/components/schemas/StructType'
        - $ref: '#/components/schemas/ListType'
        - $ref: '#/components/schemas/MapType'
      discriminator:
        propertyName: type
```

This way the OpenAPI spec also defines how to
distinguish the different types.