Python: Add pyupgrade as a pre-commit hook
The pyupgrade plugin will check the syntax, and make sure that you're
using the latest version. For example, instead of Union[str, int]
we have the str | int notation in Python 3.8.

This hook will not change any behaviour of the code, but will just
make sure that we're all using the same syntax.

More information on:
https://github.com/asottile/pyupgrade