Python: Disable autouse of the fixtures
When debugging code, the fixtures get initialized when you run
a tests in PyCharm. This is a bit anoying since it will also
hit all of your breakpoints that you have in the code.

I don't think we really need to set those on autouse, we can just
initialize on the first time we use them.