Build: Quicker project evaluation + write evaluated project version to build/
Moves the call to `getProjectVersion()` out of the "loop" that's evaluated for
every Gradle project and memoize the project version in the file `build/version.txt`.
The latter helps other projects + CI environments to inquire the version being
build or the version that has been build by looking into that file.