API: Support composite types in Accessors
Positional accessors can be helpful to getting Row values
that are not primitive. Even though there does not appear
to be any use cases for this in the Iceberg codebase itself,
it seems logical to extends Accessors to support this to
allow external tools to use Accessors uniformly across all
column types.

* Update PositionAccessor to use appropriate javaClass in
  get operations targeting composite values types (List, Map,
  Struct).

* Generate individual positional accessors for Struct fields
  to allow getting them as an Object. Previously, nested structs
  could only be accessed field-by-field.

Note: Accessing Struct fields nested in Lists and Maps is
not supported.