API: Add java type to composite TypeID enums
Previously TypeID.STRUCT, LIST and MAP exposed Void as their
java type. It seems reasonable to expose more specific java
types according to existing code patterns in Accessors and
Comparators and simplify their if/switch statements.

ManifestFileUtil has non-trivial logic around the value returned
from TypeID.javaClass() but it works only on PrimitiveType, so it
is not affected by this change.

BaseDataReader (for Spark) also has non-trivial logic for processing
TypeID.javaClass() in its convertConstant() method, but this change
appears to actually improve the processing of values in that case.

PartitionSpec appears to be unaffected by this change as it would
normally use TypeID.javaClass() only for primitive values. In any
case this change should not break its processing logic.