API: Deprecate Counter#count() / Add Counter#value()
The rationale behind this change is that usually a Counter should always
be able to report the actual count it is counting. The reason
`count()` was made `Optional` initially is because in the case of Hadoop
we are reporting the count to Hadoop's file system statistics and don't
necessarily want to report the count in that case.