API: Deprecate generic Counter<T extends Number> and replace with simpler Counter API
Most of the time using a non-generic Counter API with a long value is
simpler, especially when it comes down to serializing/deserializing
Counter results (since we'd need to know the underlying generic type as
well).

Therefore we'd want to move away from `Counter<T extends Number>` and
replace it with `Counter`.