API/Core: Don't track name in Counter/Timer API
´name()` was added to the Counter/Timer API in order to make their
results serializable. In #5427 we realized that we don't necessarily
need to track the name of a particular metric in order to serialize
counter/timer results, thus we're simplifying the Counter/Timer API by
removing `name()`.

Looking at other metric frameworks, they don't track the name of a
metric within the counter/timer itself, but rather in the metric
registry.

In our case this metric registry would be `DefaultMetricsContext` or any
other subclass that implements `MetricsContext`, so tracking the name of
a metric should be the responsibility of this registry rather than the
actual metric itself.