WiP: Make the number of metadata refresh retries configurable
For some catalog implementations the number of retries for the metadata
refresh is hard-coded to 20. For more flexibility this patch adds
support for adjusting this providing a catalog level property.

TODO: This patch covers JDBCCatalog only. Other Catalog implementations
should be taken care.
TODO: This patch doesn't have tests so far.