Core,Hive-metastore: Table owner can be incorrect in HMS
When HiveCatalog is used for creating a table the table owner property
is taken from a system property of the process running Iceberg. However,
in some clients (e.g. Apache Impala) it's not the owner of the
process who runs the create table operation and as a result not the
desired owner gets written to HMS.
For instance, in Impala's case the owner of Catalogd (the process that
makes the Iceberg API calls) was set as the table owner and not the
user who ran the CREATE TABLE statement.

I changed Hive Catalog to accept an owner via a table property from the
clients.
Note, other Catalog implementations aren't affected because the issue
is related to HMS only.