Core: Make testEnvironmentSubstitution effective when USER is not set
Make testEnvironmentSubstitution exercise the code under test even when
the USER env. var. is not set by choosing some other available env. var.

When no env. variables are available (highly unlikely) the test is ignored.

PR #5353 allowed testEnvironmentSubstitution to pass when the USER env. var.
was not set in the build environment. However, in that situation the test's
logic would be equivalent to another existing test case:
testEnvironmentSubstitutionWithMissingVar, and, more importantly, it would
not validate that the code under test was using the right environment map
(i.e. `System.getenv()`).

While the single parameter `resolveAll` method is trivial, this change allows
it to be validated even when USER is not set.