Build: Apply spotless on integration modules as well
This was unintentionally forgotten and usually spotless is able to infer
all java files from the source sets. However, the way we defined the
source sets doesn't seem working with spotless and so
we were defining relevant modules via `target`.

Additional details:

Spotless requires all source sets to be part of the current module being
evaluated, which isn't the case for the hive module:
```
> Spotless error! All target files must be within the project dir.
    project dir: /home/nastra/Development/workspace/iceberg/hive3
         target: /home/nastra/Development/workspace/iceberg/mr/src/test/java/org/apache/iceberg/mr/hive/HiveIcebergStorageHandlerTestUtils.java
```

And this is how the sourceSets are defined in `iceberg-hive3` which
cause issues:

```
project(':iceberg-hive3') {

  // run the tests in iceberg-mr with Hive3 dependencies
  sourceSets {
    test {
      java.srcDirs = ['../mr/src/test/java', 'src/test/java']
      resources.srcDirs = ['../mr/src/test/resources', 'src/test/resources']
    }
  }
...
}
```