Python: Test if version is PEP440 compliant
Turned out that my regex isn't PEP440 compatible.

RCs are represented as 0.1.0rc1 and dev versions are with a dot: 0.1.0.dev1
And, they can be combined: 0.1.0rc1.dev1.

Instead of figuring out the regex myself, we can just use the one from packaging.

This package is already installed because it is also a dependency of pytest.