Python: Catch ValidationError on invalid REST response
If an implementation of the rest catalog returns something different than
we expect, we just propagate the ValidationError, which subclasses the
ValueError. When using the CLI, it will tell you that the URI is not set,
but actually it is, it is just invalid.

This will return the json to the user, and also tell what's wrong with the
response.