Python: Use Types from Typing
Instead of using the runtime types that are available using the futures
import:

```python
from __future__ import annotations
```

I prefer to use the types from Typing:

- I think `Optional[str]` is clearer than `str | None`.
- Not all types are available, so we have to import from Types anyway (typevar, Callable, etc)

PyUpgrade was rewriting the [imports automatically](https://github.com/asottile/pyupgrade#pep-585-typing-rewrites),
so we ended up with a mix, but I think it is better to stay
consistent. I've disabled this on PyUpgrade.