Python: Fix Evaluator tests
Instead of just supplying the unbound expressions,
we replaced the bound expression, with the one that we wanted
to test.
But this caused issues with that we only used a long schema,
so all the string and double literals were casted to a LongLiteral.

This PR changes the tests to supply the correct schema, and includes
the binding in the tests as well