Python: Improve docs
This improves the docs by turning into a multi page site.

Also merges in the `CONTRIBUTING.md` and `RELEASE.md` to have
everything into a single place.

Also changed to the readthedocs theme since that has a much
cleaner multi page layout (navigation on the left instead
of the top).