Python-legacy: Fix CI
It looks like the latest version of `tox` clashes with `tox-travis`.

The tox version wasn't pinned in `tox-travis`, and it would just
follow the latest version that's available.

This has been fixed in the repository:
https://github.com/tox-dev/tox-travis/commit/e11bacf107903375b5c42bda799e94f5dbe12978

I think it is good to add this constraint in our `setup.py` as well.