Parquet: Fixes Incorrect Skipping of RowGroups with NaNs
Previously, the RowGroupFilter would skip any RowGroups whose min or max value was set as NaN because Parquet would report that those row groups "hasNonNull" as false. When the statistics are not empty and hasNonNull is false this actually indicates that the min and max are undefined, not that there are no non nulls.