API, Spark: Use Bulk Delete Methods instead of Single Deletes when Available
Previously deletes were handled by a per Action execution service that would be used
to parallelize single deletes. In this PR we move the responsibility of performing the
deletes and the parallelization of those deletes to the FileIO via SupportsBulkOperations.

This deprecates all methods which used to be used for doing single deletes as well as
passing executor services to Actions which delete many files.