Python: Fix warnings from PyLint
Because TestType contains test in the name,
pytest tries to collect it and run tests on it.
Setting `__test__ = False` tells PyTest to ignore it

Also, s3 emits an error because we clean up buckets that
still has content in it, so I've added some code to clean them
up after we use the bucket.