Python: Pass location when recreating the FileIO
We need to pass in the metadata location when we recreate
the file io for the table. We re-create the table because
there can be new configuration, but we don't supply the location.

This can cause issues when you have both PyArrow and FSspec
installed, you don't have the FileIO set explicitly in your config:

```yaml
catalog:
    default:
...
        py-io-impl: pyiceberg.io.fsspec.FsspecFileIO
```

When you fetch the metadata from ADLS using FSSpec,
but then the re-created FileIO is PyArrow.
When we pass in the url, the schema gets taken into
account and PyArrow won't be considered for ADLS.