Make `TableMetadataParser.fromJson` taking `JsonNode` public
In the WIP for the Nessie Iceberg REST catalog server, there are a few places that currently require us to first serialize a `TableMetadata` to a `String` just to parse it right back into a `TableMetadata`, but this time with the right metadata-location.

First occurence is when we write the table metadata, where we have to first store the table metadata, get the metadata location, and then return the table metadata with the metadata location (and updates cleared).
Second occurence is when we return a table metadata with "transient properties", which are used to pass the "expected Nessie commit" around.

This change makes the two `TableMetadataParser.fromJson()` taking a `JsonNode` public