Create gradle.yml
Honestly, I dunno what I am doing.  But I'm trying and struggling. There are times when the code won't operate. I then dig in to see why... whether I didn't setup correctly or the code is corrupt. Which it is. So many different commands and code. Learning not to just download a repository and hope it works... duh. What I need to know is the how's and why's. I have been going through these difficulties for years. Not something that is just "new". I'm exhausted and I wanna tap out. But there is defiance within me and I do not have the heart to just walk away. I call BS. I am aware that there are specific people who would harass me as such.

Attempt #290 -hoping the breaks the cycle