Spec: Make `iceberg.schema` part of the spec
I noticed that the Iceberg Java writers are setting this field
on the Parquet/Avro/ORC files, but others (I can confirm that
this is missing for Athena and Trino) don't do this.

I think this is nice because we just hit an edge case:
https://github.com/apache/iceberg/pull/7523

Where the Arrow reader doesn't properly infer the
logical UUID type. With the schema in there, we know
which schema that was used for writing the file,
and we can use that as the source of truth.