Data: TableMigrationUtil shouldn't use path to get partition values
The existing implementation of TableMigrationUtil.loadPartition() uses the partition path to guess
the partition values. However, if a string partition column contains values that require
URL-encoding then the partition values will be wrong.
For instance the string value "11/12/13" on the partition path is URL-encoded to '11%2F12%2F13'. If
you run loadPartition() then the resulting DataFiles will incorrectly have the URL-encoded
partition values.

As a solution this patch adds a new implementation of loadPartition() that doesn't require the
caller to pass the partition values in a URL-encoded format and creates DataFiles having the
correct partition values.

Note, I kept the existing implementation as well and marked it as deprecated so that it can be
dropped in a later release.