Remove the direct dependency on Apache Commons.
The approach followed is to use native Java 8 code when the alternative was clear or replicate/borrow when not following iceberg-core project practice.

Flink and Spark projects declare Apache Commons as direct dependency and therefore on those subprojects the dependency is not removed.

Remove the direct dependency on Apache Commons.

The approach followed is to use native Java 8 code when the alternative was clear or replicate/borrow when not following iceberg-core project practice.

Flink and Spark projects declare Apache Commons as direct dependency and therefore on those subprojects the dependency is not removed.

Remove the direct dependency on Apache Commons.

The approach followed is to use native Java 8 code when the alternative was clear or replicate/borrow when not following iceberg-core project practice.

Flink and Spark projects declare Apache Commons as direct dependency and therefore on those subprojects the dependency is not removed.