Python: improve Avro parsing by caching struct types.
Cache the struct types for ManifestEntry and ManifestFile schemas.

Cache the _position_to_field_name lookup in Records if a struct type is passed.

Improve StructReader to use slots and avoid isinstance() lookups on each time
the structure is read, by performing the check in __init__().

Improve StructReader to determine the arguments to create_struct() once in init
and reuse the result each time the structure is read, avoid a try block.