Python: Improve Avro Parse speed
Cache the hash value of the Struct types so that when new
Record objects are created time is not spent recomputing
the hash value.

Change the implementation of the Record class to cache the
mapping from field name to position.  Also utilize an array
rather than a dictionary for these lookups.