Move validations into Builder to be more lenient when ViewMetadata is read through Parser
There are currently only two ways to create an instance of
`ViewMetadata`:
* through the Builder
* through the Parser

Almost all validations should be happening in the Builder, while we
would like to be more lenient when view metadata is being read (e.g.
from disk) via the parser.

Reading view metadata through the parser, it is acceptable that things
like schemas/versions/history might be empty or that the current version
id points to a version that doesn't exist. In such a case we'd rather
want to lazily fail when accessing the view's current version rather
than when the view metadata is being read by the parser.