Docs: Clarified Partition Transform
As an Iceberg newbie, I was very confused by the difference between the partitioning overview and the partition transform spec documentation.  The partitioning overview page seemed to reference a transform named "date".  However, this transform does not actually exist.  Instead "date" was intended to be a pseudo transform demonstrating the possibilities of Iceberg's partition transforms.

This use of "date" pseudo transform resulted in confusion for me as I could not find any such partition transform.  After digging into the spec docs more, I realized the "day" partition transform is the same as this "date" pseudo transform.

This commit updates the doc to use the actual "day" transform in the example.