Core: Fix missing delete files from transaction
With retries with conflicting manifest merges.

Ryan pointed out that this might also occur whith the deletes.
However, I was unable to replicate this with a test. I've added
the test that should uncover this issue when merging DELETE
manifests, and deleting the old one before the transaction
is succesfully commited.