Build: Define strict version for Flink / Jackson / Hive2 / Tez 0.8
When running into this issue, the build used Flink 1.18.0.
I've noticed that we were silently using Flink 1.18.1 because a new patch version was just released, which caused a test to fail.
This is because of how `strictly` and `prefer` (or even `require`)
version definitions typically work.
According https://docs.gradle.org/current/userguide/rich_versions.html
Gradle would select a patch version as long as it matches the version
range defined in `strictly`, which isn't necessarily what we want for
those libraries.
I've changed them all to not use version ranges.