Core: RowDelta should conflict with RewriteFiles operation
RewriteFiles might be used to compact a table, e.g. it can replace
lots of small files with a few big files. Concurrent RowDelta (DELETE/
UPDATE) operations must fail, because they can put the table in an
inconsistent state.

A successful DELETE can become a no-op as the delete files refer to
the replaced small files.

A successful UPDATE (with merge-on-read) can duplicate rows as the
delete files refer to the replaced small files, and it also adds
its own data files.