Fix `rewrite_position_delete_files` result file set
The SparkBinPackPositionDeletesRewriter module disables Adaptive Query Execution (AQE) to correct issues with partitioning.

However, if the parent Spark session is configured with a `spark.sql.shuffle.partitions` value not equal to 1, this can lead to the creation of multiple delete files, rather than a single file.

To address this issue, enforce setting sql.shuffle.partitions to 1.