AWS: Use hadoop bulk delete API where available
Reflection-based used of Hadoop 3.4.1+ BulkDelete API so that
S3 object deletions can be done in pages of objects, rather
than one at a time.

Configuration option "iceberg.hadoop.bulk.delete.enabled" to switch
to bulk deletes.

There's a unit test which will turn this on if the wrapped APIs
are loaded and probe the HadoopFileIO instance for it using
the APIs.

* Parameterized tests for bulk delete on/off
* Cache bulk delete page size; this brings performance of bulk delete
   with page size == 1 to that of single delete
* Use deleteFiles() in tests which create many files; helps highlight
  performance/scale issues against local fs.

+ updates aws.md to cover this and other relevant s3a settings,
  primarily for maximum parquet IO