HADOOP-18679. PoC of using bulk delete API in iceberg
Highlights that we need to support bulk delete on all FS instances,
because there seems to be no guarantee that every path
is from the same FS.

Certainly the S3FileIO impl supports multiple buckets,
which is where the code to gather paths by bucket and schedule comes
from.

Yes, it's a bit over-complex for single file filesystems, but
no obvious inefficiencies.

Not actually compiled/tested...not sure how to switch profiles