Don't let `o.a.i.util.Tasks` log unnecessary stack traces
When retries happen, each retry produces a full stack trace on the console (or log file), which is, in the vast majority of all cases, just not needed and unnecessarily confuses users, because users assume that something went terribly wrong, when they see a stack trace.

This change removes the logging of the stack trace from the normal "retry warning", which is still logged at "warning" level. If debug logging is enables, those logs will emit the stack trace.