Fix issues.
- When converting an Iceberg filter to a Parquet filter, and then using
the converted filter in Parquet to filter row groups, Parquet checks that
the type of a column in a filter predicate matches the type in the Parquet
file as a validation step before applying the filter. This validation fails
for some cases, e.g., INT96 timestamp columns. When doing the conversion,
we thus need to check that such a type mismatch does not occur and fail
the conversion if it does.
- When converting the Iceberg filter to a Parquet filter fails, we need
to handle the failure, and in ReadConf, we need to use the internally
computed total number of rows instead of the values returned by the
ParquetFileReader's getFilteredRecordCount().
- In ParquetReader.FileIterator, since we have to handle both cases where
a Parquet record filter is used and where it is not, we avoid the
skipNextRowGroup() and readNextFilteredRowGroup() methods of ParquetFileReader
and instead proceed row group by row group and call readFilteredRowGroup(int)
with the index of the row group.